requirements.txt
discord.py-self
pyyaml
python-dotenv
apscheduler
pytz
